# apps/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.AssignRolesRequest:
    properties:
      roleIds:
        items:
          type: integer
        type: array
    required:
    - roleIds
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateUserRequest:
    properties:
      deptId:
        type: integer
      email:
        maxLength: 128
        type: string
      gender:
        enum:
        - 0
        - 1
        - 2
        type: integer
      nickname:
        maxLength: 64
        type: string
      password:
        maxLength: 64
        minLength: 6
        type: string
      phone:
        type: string
      remark:
        maxLength: 255
        type: string
      roleIds:
        items:
          type: integer
        type: array
      status:
        enum:
        - 0
        - 1
        type: integer
      username:
        maxLength: 64
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.ResetPasswordRequest:
    properties:
      password:
        maxLength: 64
        minLength: 6
        type: string
    required:
    - password
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateProfileRequest:
    properties:
      email:
        maxLength: 128
        type: string
      gender:
        enum:
        - 0
        - 1
        - 2
        type: integer
      nickname:
        maxLength: 64
        type: string
      phone:
        type: string
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateUserRequest:
    properties:
      deptId:
        type: integer
      email:
        maxLength: 128
        type: string
      gender:
        enum:
        - 0
        - 1
        - 2
        type: integer
      nickname:
        maxLength: 64
        type: string
      phone:
        type: string
      remark:
        maxLength: 255
        type: string
      roleIds:
        items:
          type: integer
        type: array
      status:
        enum:
        - 0
        - 1
        type: integer
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.RoleItem:
    properties:
      code:
        type: string
      createdAt:
        type: string
      dataScope:
        type: integer
      id:
        type: integer
      name:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: integer
    type: object
...
  github_com_fangzhengcn_workbackend_apps_backend_internal_vo.UserItem:
    properties:
      createdAt:
        type: string
      deptId:
        type: integer
      deptName:
        type: string
      email:
        type: string
      gender:
        type: integer
      id:
        type: integer
      lastLoginAt:
        type: string
      nickname:
        type: string
      phone:
        type: string
      remark:
        type: string
      roles:
        items:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.RoleItem'
        type: array
      status:
        type: integer
      username:
        type: string
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData:
    properties:
      list: {}
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: 基于 RBAC 的通用后台管理系统接口
  title: 权限管理后台 API
  version: "1.0"
paths:
  /auth/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 头像图片（JPG/PNG/GIF/WEBP，不超过 2MB）
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 头像访问 URL
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  type: string
              type: object
      summary: 上传头像
      tags:
      - 个人中心
...
  /auth/profile:
    put:
      consumes:
      - application/json
      parameters:
      - description: 个人资料
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 修改个人资料
      tags:
      - 个人中心
...
  /users:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.PageData'
              type: object
      summary: 用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 新增用户
      tags:
      - 用户管理
  /users/{id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 删除用户
      tags:
      - 用户管理
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_vo.UserItem'
              type: object
      summary: 用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 修改用户
      tags:
      - 用户管理
  /users/{id}/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 密码参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 重置密码
      tags:
      - 用户管理
  /users/{id}/roles:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_internal_dto.AssignRolesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fangzhengcn_workbackend_apps_backend_pkg_response.Body'
      summary: 分配角色
      tags:
      - 用户管理
  /users/export:
    get:
      parameters:
      - in: query
        name: deptId
        type: integer
      - in: query
        minimum: 1
        name: page
        type: integer
      - description: Phone 只能精确匹配：手机号为密文，需转成 HMAC 盲索引后查询。
        in: query
        maxLength: 32
        name: phone
        type: string
      - in: query
        maximum: 200
        minimum: 1
        name: size
        type: integer
      - enum:
        - 0
        - 1
        in: query
        name: status
        type: integer
      - description: Username 支持模糊查询（明文列）。
        in: query
        maxLength: 64
        name: username
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: CSV 文件
          schema:
            type: file
      summary: 导出用户
      tags:
      - 用户管理
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"